LICENSE.md: remove misleading phrasing in preamble
The license is not a copyleft license as it does not grant the four
freedoms.  In particular, it restricts distribution making it
a proprietary license better referred to as source-available license.

Fun fact: by creating this pull request I’ve technically violated the
license since I had to create fork of the software.